chore(plot): move CallbackSchedRectSource and CallbackSchedBarSource into callback_scheduling_source.py
