+++ src/caret_analyze/plot/visualize_lib/bokeh/bokeh_source.py
@@ -16,16 +16,11 @@ from __future__ import annotations
 
 from collections import defaultdict
 from logging import getLogger
-from typing import Any, Dict, List, Optional, Tuple, Union
+from typing import Any, Dict, List, Tuple, Union
 
 from bokeh.models import GlyphRenderer, HoverTool, Legend
-from bokeh.plotting import ColumnDataSource
 
-from .util import RectValues
-
-from ....common import ClockConverter
 from ....exceptions import InvalidArgumentError
-from ....record import Clip
 from ....runtime import (CallbackBase, Communication, Publisher, Subscription,
                          SubscriptionCallback, TimerCallback)
 
@@ -175,155 +170,6 @@ class LegendSource:
             raise NotImplementedError()
 
         return description
-
-
-class CallbackSchedRectSource:
-    """Class to generate callback scheduling rect sources."""
-
-    RECT_HEIGHT = 0.3
-    _RECT_Y_STEP = -1.5
-
-    def __init__(
-        self,
-        legend_manager: LegendManager,
-        target_object: TimeSeriesTypes,
-        clip: Clip,
-        converter: Optional[ClockConverter] = None
-    ) -> None:
-        self._legend_keys = LegendKeys('callback_scheduling_rect', target_object)
-        self._hover = HoverCreator(self._legend_keys)
-        self._legend_source = LegendSource(legend_manager, self._legend_keys)
-        self._clip = clip
-        self._converter = converter
-        self._rect_y_base = 0.0
-
-    @property
-    def rect_y_base(self) -> float:
-        return self._rect_y_base
-
-    def create_hover(self, options: dict = {}) -> HoverTool:
-        """
-        Create HoverTool based on the legend keys.
-
-        Parameters
-        ----------
-        options : dict, optional
-            Additional options, by default {}
-
-        Returns
-        -------
-        HoverTool
-
-        """
-        return self._hover.create(options)
-
-    def generate(self, callback: CallbackBase) -> ColumnDataSource:
-        """
-        Generate callback scheduling rect source.
-
-        Parameters
-        ----------
-        callback : CallbackBase
-            target callback.
-
-        Returns
-        -------
-        ColumnDataSource
-
-        """
-        rect_source = ColumnDataSource(data={
-            k: [] for k in (['x', 'y', 'width', 'height'] + self._legend_keys.to_list())
-        })
-        latency_table = callback.to_dataframe(shaper=self._clip)
-        for row in latency_table.itertuples():
-            callback_start = self._converter.convert(row[1]) if self._converter else row[1]
-            callback_end = self._converter.convert(row[-1]) if self._converter else row[-1]
-            rect = RectValues(
-                callback_start, callback_end,
-                (self._rect_y_base-self.RECT_HEIGHT),
-                (self._rect_y_base+self.RECT_HEIGHT)
-            )
-            rect_source.stream({
-                'legend_label': [
-                    self._legend_source.get_non_property_data(callback, 'legend_label')
-                ],
-                'x': [rect.x],
-                'y': [rect.y],
-                'width': [rect.width],
-                'height': [rect.height],
-                'callback_start': [f'callback_start = {callback_start} [ns]'],
-                'callback_end': [f'callback_end = {callback_end} [ns]'],
-                'latency': [f'latency = {(callback_end - callback_start) * 1.0e-6} [ms]']
-            })
-
-        return rect_source
-
-    def update_rect_y_base(self) -> None:
-        """Update rect_y_base to the next step."""
-        self._rect_y_base += self._RECT_Y_STEP
-
-
-class CallbackSchedBarSource:
-    """Class to generate callback scheduling bar sources."""
-
-    def __init__(
-        self,
-        legend_manager: LegendManager,
-        target_object: TimeSeriesTypes,
-        frame_min: float,
-        frame_max: float
-    ) -> None:
-        self._legend_keys = LegendKeys('callback_scheduling_bar', target_object)
-        self._hover = HoverCreator(self._legend_keys)
-        self._legend_source = LegendSource(legend_manager, self._legend_keys)
-        self._frame_min = frame_min
-        self._frame_max = frame_max
-
-    def create_hover(self, options: dict = {}) -> HoverTool:
-        """
-        Create HoverTool based on the legend keys.
-
-        Parameters
-        ----------
-        options : dict, optional
-            Additional options, by default {}
-
-        Returns
-        -------
-        HoverTool
-
-        """
-        return self._hover.create(options)
-
-    def generate(self, callback: CallbackBase, rect_y_base: float) -> ColumnDataSource:
-        """
-        Generate callback scheduling bar source.
-
-        Parameters
-        ----------
-        callback : CallbackBase
-            target callback.
-        rect_y_base : float
-            The y-base of rect.
-
-        Returns
-        -------
-        ColumnDataSource
-
-        """
-        rect = RectValues(
-            self._frame_min, self._frame_max,
-            rect_y_base - 0.5,
-            rect_y_base + 0.5
-        )
-        legend_source = self._legend_source.generate(callback)
-        bar_source = ColumnDataSource(data={
-            **{'x': [rect.x], 'y': [rect.y],
-               'width': [rect.width], 'height': [rect.height]},
-            **legend_source  # type: ignore
-        })
-
-        return bar_source
 
 
 class LegendManager:
